% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_same.R
\name{all_same}
\alias{all_same}
\title{All Values Same}
\usage{
all_same(x)
}
\arguments{
\item{x}{A vector of values. Can be one of the following objects
classes: \code{numeric}, \code{character}, \code{factor}, or \code{logical}.}
}
\value{
Logical. \code{TRUE} or \code{FALSE}.
}
\description{
Returns a logical as to whether all values \emph{within} a vector
are identical. This function does NOT compare two independent
vectors. Please use \code{\link[=all.equal]{all.equal()}}, for such a purpose
combined with \code{\link[=isTRUE]{isTRUE()}}.
}
\examples{
all_same(1:4)
all_same(rep(5, 10))
all_same(rep("A", 10))
all_same(letters)
all_same(c(TRUE, TRUE, TRUE))
all_same(c(TRUE, TRUE, FALSE))
}
\seealso{
\code{\link[=isTRUE]{isTRUE()}}, \code{\link[=all.equal]{all.equal()}}
}
\author{
Stu Field
}
