% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignRData.R
\name{assignRData}
\alias{assignRData}
\title{Assign Rdata Contents to a Variable}
\usage{
assignRData(file)
}
\arguments{
\item{file}{The character path to an \code{*.Rdata} or \code{*.rda} file.}
}
\value{
If the binary Rdata file only contains a single object then
this object is returned. If the file contains more than one object,
a named list is returned.
}
\description{
Assigns the contents of a binary \code{*.Rdata} or \code{*.rda}
file to a variable rather than loading it directly into the
global environment. Useful for avoiding unexpected collisions
with variables in the current global environment.
}
\examples{

\dontrun{
assignRData("path/to/myfile.rda")
}

}
\seealso{
\code{\link{load}}.
}
\author{
Mike Mehan
}
