% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMatrixSensitivity.R
\name{calcMatrixSensitivity}
\alias{calcMatrixSensitivity}
\title{Linear Sensitivity Analysis}
\usage{
calcMatrixSensitivity(A)
}
\arguments{
\item{A}{A projection matrix to be analyzed.}
}
\value{
A list of the following:
\item{Matrix }{The original projection matrix (map)}
\item{Sensitivity }{Matrix of linear sensitivity based on
perturbations of parameters to lambda}
\item{Elasticity }{Matrix of linear elasticity (proportional change)
based on perturbations of parameters to lambda}
\item{lambda }{The dominant eigenvalue of A. Projection matrix population
"growth"}
\item{w }{The right eigenvector or stable stage distribution of the
population}
\item{v }{The left eigenvector or the reproductive "value" of each stage
in the population}
}
\description{
Sensitivity analysis of linear maps, all in one function.
}
\examples{
A <- diag(1:5 / 10)
A[cbind(2:5, 1:4)] <- 3:6 / 10
A[1, 5] <- 5

calcMatrixSensitivity(A)
}
\references{
Caswell, H. Matrix Population Models. 2001. Sensitivity & Elasticity.
}
\seealso{
\code{\link{analyzeProjectionMatrix}}, \code{\link{eigen}},
\code{\link[popbio]{eigen.analysis}}, \code{\link[popbio]{popbio}}
}
\author{
Stu Field
}
