% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capwords.R
\name{capwords}
\alias{capwords}
\title{Capitalize Title Format}
\usage{
capwords(s, strict = FALSE)
}
\arguments{
\item{s}{A character string in the form of a sentence to be converted to
"title case" (i.e. first letter capitalized).}

\item{strict}{Logical. Should first letter capitalization be \emph{strictly}
applied? See example.}
}
\value{
A character string with "title case" conversion.
}
\description{
Change case to capitalize first letter of each word in a character string.
}
\examples{
capwords("using AIC for model selection")
## ->  [1] "Using AIC For Model Selection"
capwords(c("using AIC", "for MODEL selection"), strict = TRUE)
## ->  [1] "Using Aic"  "For Model Selection"
##                ^^^        ^^^^^
##               "bad"       "good"
}
\seealso{
\code{\link[base]{toupper}}, \code{\link[base]{tolower}}, \code{\link[base]{strsplit}}
}
\author{
Stu Field
}
