% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI95se.R
\name{CI95se}
\alias{CI95se}
\title{Calculate SEM 95\% Confidence Intervals}
\usage{
CI95se(x)
}
\arguments{
\item{x}{Numeric. A vector of data to calculate the CI95.}
}
\value{
A numeric vector with 3 named entries:
\item{lower}{the lower CI95.}
\item{mean}{the arithmetic mean.}
\item{upper}{the upper CI95.}
}
\description{
Uses standard error (of the mean) calculation to
determine the 95\% CI of a vector of data.
Does not use a bootstrapping of empirical data,
but the Gaussian approximation.
}
\examples{
CI95se(rnorm(100))
}
\seealso{
\code{\link[=bootstrap]{bootstrap()}} for CI95 via bootstrapping.
}
\author{
Stu Field
}
