% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefRi.R
\name{coefRi}
\alias{coefRi}
\title{Calculate Interclass Correlation Coefficient (ICC)}
\usage{
coefRi(x, groups)
}
\arguments{
\item{x}{A matrix or data frame containing the raw data of the various
treatments you are testing in the ANOVA.}

\item{groups}{A vector of the factor groupings for \code{x}.}
}
\value{
A list containing:
\item{model}{Resulting ANOVA table}
\item{ICC}{The intraclass correlation coefficient, the measure of
similarity among individuals \emph{within} a treatment
group relative to the differences found \emph{among} groups}
}
\description{
The Intraclass correlation coefficient (r_i; aka ICC) can be used to
estimate the repeatability of a method. The value 0 -> 1. Depending on how
the groups are set up, you want all your variation to be among groups
(individuals), not within groups (repeats) so you want this value to be high
if individuals are your groups, and low if your repeated measurements are
the groups. When ICC is high, it means most of the variation is
\emph{between} treatment groups.
}
\examples{
head(Ri_data)    # internal data
x <- as.vector(as.matrix(Ri_data))[ !is.na(as.vector(as.matrix(Ri_data))) ]
coefRi(x, groups = rep(names(Ri_data), c(8, 10, 13, 6)))
}
\references{
Sokal & Rohlf (Biometry; 3rd ed.) 210-214. Sokal & Rohlf
(Biometry; 2rd ed.) 211-216.
}
\seealso{
\code{\link[=aov]{aov()}}
}
\author{
Stu Field
}
