% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse2df.R
\name{collapse2df}
\alias{collapse2df}
\title{Collapse Vector List to Data Frame}
\usage{
collapse2df(x)
}
\arguments{
\item{x}{A \emph{named} list (which become the rows) containing numeric
vectors of the same length. Vectors can be in the form of a numeric vector
or a 1-row data frame}
}
\value{
A data frame object of the vertically collapsed vectors
}
\description{
Function collapses a list of vectors, or a list of 1-row data frames, into a
n x m data frame, where n equals the list length and m equals the vector
length. All vectors must be of the same length.
}
\details{
Uses \code{\link[base]{rbind}} and \code{\link[base]{Reduce}} and to preform
the collapse.
}
\note{
Functions similarly to \code{\link[base]{do.call}}.
}
\examples{
tmp <- lapply(1:3, function(...) rnorm(4))
names(tmp) <- head(LETTERS, 3)
collapse2df(tmp)
tmp2 <- lapply(tmp, function(x) { names(x) <- head(letters, 4); x })
collapse2df(tmp2)
}
\seealso{
\code{\link[base]{rbind}}, \code{\link[base]{Reduce}}
}
\author{
Stu Field
}
