% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_tabulate.R
\name{cross_tabulate}
\alias{cross_tabulate}
\title{Cross Tabulate Counts}
\usage{
cross_tabulate(x, ...)
}
\arguments{
\item{x}{A \code{soma_adat} or \code{data.frame} object containing the
meta data from which counts are desired.}

\item{...}{A number of un-quoted expressions separated by commas. Generally
greater than 3 is un-useful.}
}
\value{
A table of grouped counts based on splitting
variables with sums from each factor.
}
\description{
Create a contingency table of counts generated by cross-classifying
factors from groups splitting on the values passed in the \code{...} argument.
The sums of each row and column are added to the result.
}
\examples{
# 1 factor
cross_tabulate(mtcars, cyl)
# 2 factors
cross_tabulate(mtcars, cyl, gear)
# 3 factors
cross_tabulate(mtcars, cyl, gear, am)
}
\seealso{
\code{\link[=table]{table()}}, \code{\link[=addmargins]{addmargins()}}
}
\author{
Stu Field
}
