% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossTab.R
\name{crossTab}
\alias{crossTab}
\alias{my.table}
\title{Cross Tabulate Summary Counts}
\usage{
crossTab(x, by)
}
\arguments{
\item{x}{A "data.frame" or "tibble" object containing the
data from which counts are desired.}

\item{by}{Character. The grouping variable(s). Can be of
\code{length=1} or \code{length=2}, if \code{length=2} a 2-dim
table will be returned.}
}
\value{
A table of grouped counts based on splitting variables with sums
from each factor.
}
\description{
Create a contingency table of counts generated by cross-classifying
factors from groups splitting on the \code{by=} argument,
and an optional secondary splitting variable. The sums of each
row and column are added to the result.
}
\note{
This is a simple wrapper around \code{\link{table}}
that adds the sums of columns and rows to the final object
}
\examples{

crossTab(test_data, by = "Sample")
crossTab(test_data, by = c("Sample", "TimePoint"))

}
\seealso{
\code{\link{table}}, \code{\link[stats]{addmargins}}
}
\author{
Stu Field
}
