% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.R
\name{cumulative}
\alias{cumulative}
\title{Cumulative Wrapper}
\usage{
cumulative(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list containing:
\item{cum_min}{The cumulative minima of the elements of \code{x}}
\item{cum_max}{The cumulative maxima of the elements of \code{x}}
\item{sum}{The sum of the elements of \code{x}}
\item{cum_sum}{The cumulative sum of the elements of \code{x}}
\item{prod}{The product of the elements of \code{x}}
\item{cum_prod}{The cumulative product of the elements of \code{x}}
}
\description{
Returns cumulative vector values of neighboring vector elements.
}
\note{
Used as a mere exercise in function writing for the R tutorial.
}
\examples{
\dontrun{
?cumsum
}
cumulative(1:10)
}
\seealso{
\code{\link{cumsum}}, \code{\link{cumprod}}
}
\author{
Stu Field
}
