% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagR.R
\name{diagR}
\alias{diagR}
\title{Create Matrix from Vector}
\usage{
diagR(x, k = 0)
}
\arguments{
\item{x}{Numeric. A Vector to be placed into entries
of a matrix along a diagonal.}

\item{k}{Numeric. The offset from the diagonal.
+1 = super-diagonal, -1 = sub-diagonal.
Default = 0, which reverts to \code{\link[=diag]{diag()}}.}
}
\value{
A matrix with vector "x" along a diagonal, or the
sub- super-diagonal if \code{k != 0}.
}
\description{
Matlab mimic function for producing matrices with vectors
along the sub- or super-diagonal.
}
\note{
Matlab style matrix \code{\link[=diag]{diag()}}
}
\examples{
diagR(1:7, k = -1)     # sub-diagonal
diagR(1:15, k = 0)     # same as diag()
vec <- seq(15, 30, by = 3)
diagR(vec, 1)          # super-diagonal
}
\seealso{
\code{\link[=diag]{diag()}}, \code{\link[=zeros]{zeros()}}
}
\author{
Stu Field
}
