% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriftSim.R
\name{DriftSim}
\alias{DriftSim}
\title{Genetic Drift Simulation}
\usage{
DriftSim(p.star = 0.5, n = 50, nsim = 50, plot = "b")
}
\arguments{
\item{p.star}{Initial allelic frequency of the p allele.}

\item{n}{Integer. Number of individuals in the population.}

\item{nsim}{Integer. Number of simulations to perform.}

\item{plot}{Plot both histogram and drift with time (default) or just
histogram ("h") or just line graph ("l").}
}
\value{
A plot of the simulation is returned.
}
\description{
Performs an illustrative simulation of genetic drift.
}
\note{
Simulation written for FEScUE class
}
\examples{
\dontrun{
DriftSim()
}
}
\references{
Department of Biology, Colorado State University,
Fort Collins, CO 80523-1878.
}
\seealso{
\code{\link{DriftSim2}}
}
\author{
Stu Field
}
