% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriftSim2.R
\name{DriftSim2}
\alias{DriftSim2}
\title{Genetic Drift Simulation 2}
\usage{
DriftSim2(p, Gen, n, trials)
}
\arguments{
\item{p}{Numeric. Initial allelic frequency in for each of the simulations}

\item{Gen}{Integer. Generations to run each simulation}

\item{n}{Integer. Number of individuals in each of the trials}

\item{trials}{Integer. Number of simulations to run}
}
\value{
A list containing:
\item{p.mat }{A matrix composed of the simulation of each of the
drift simulations following the allelic frequency p.}
\item{P.fix }{The proportion of the time the p allele becomes fixed
in the population. Should approach 0.5 when trials increases.}
}
\description{
Performs a simulation of genetic drift according to the one described in the
Evolution & Ecology Excel spreadsheets manuals
}
\examples{
\dontrun{
set.seed(501)
DriftSim2(p = 0.5, Gen = 100, n = 10, trials = 20)
}
}
\references{
Evolutionary Ecology Tutorials in Excel - Workbook.
}
\seealso{
\code{\link{DriftSim}}
}
\author{
Stu Field
}
