% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_file.R
\name{find_file}
\alias{find_file}
\title{Determine Path to File}
\usage{
find_file(pattern, root = Sys.getenv("HOME"))
}
\arguments{
\item{pattern}{Character string of the pattern match.}

\item{root}{The path of the root directory to start the highest level
search.}
}
\value{
The path(s) containing the regular expression specified in \code{pattern}.
}
\description{
Determine the path of a defined file via brute force search of
(optionally) the root directory.
}
\examples{
\dontrun{
find_file("infection_data.csv")
fs::file_create("myfile.txt")
find_file("myfile.txt")
}
}
