% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileExt.R
\name{getFileExt}
\alias{getFileExt}
\title{Get File Extension}
\usage{
getFileExt(x)
}
\arguments{
\item{x}{Character. String of the file name. Can be absolute or relative
path, or just the basename.}
}
\value{
A string containing the extension of the file name passed in
\code{x}.
}
\description{
Function to return the file extension of a character string given a path
name to a file.
}
\examples{

getFileExt("file.pdf")
getFileExt("/home/full/path/to/file.pdf")

}
\seealso{
\code{\link[base]{regexpr}}, \code{\link[base]{substring}}
}
\author{
Stu Field
}
