% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagged_tbl.R
\name{jagged_tbl}
\alias{jagged_tbl}
\title{Create a Jagged Tibble}
\usage{
jagged_tbl(x, col.names = NULL)
}
\arguments{
\item{x}{A list of vectors to convert to 2-dimantional \code{tibble}.}

\item{col.names}{Character. A vector of names for the resulting \code{tibble}.}
}
\value{
A \code{tibble} with columns coerced from list vector entries.
}
\description{
From a list of un-equal length vectors, fill in lengths with \code{NA} and
convert to a \code{tibble} object.
}
\examples{
# blah
}
