% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matPower.R
\name{matPower}
\alias{matPower}
\title{Matrix Power}
\usage{
matPower(X, n)
}
\arguments{
\item{X}{A matrix to be multiplied.}

\item{n}{An integer to place in the exponent (power).}
}
\value{
A matrix, the result of X^n.
}
\description{
Calculate the result of exponentiation of a matrix of
the form \deqn{X^n}, where \code{X} is a matrix and \code{n} is an integer.
}
\details{
This technique is referred to the Power Method and is sometimes
used to estimate Lambda, the growth rate of a projection matrix.
}
\examples{
M <- diag(c(1.3, 0.4, 0.5, 0.9))
matPower(M, 10)
}
\author{
Stu Field
}
