% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixX.R
\name{matrixProd}
\alias{matrixProd}
\title{Matrix Multiplication}
\usage{
matrixProd(A, B)
}
\arguments{
\item{A}{The pre-multiplied matrix.}

\item{B}{The post-multiplied matrix.}
}
\value{
A matrix (or vector) result of matrix multiplication.
}
\description{
Mimic of proper matrix multiplication without
using the R syntax of \%*\%.
}
\details{
For matrix multiplication, the ncol(A) must equal nrow(B)! and you get a
nrow(A) x ncol(B) matrix as a result. Must be matrices not vectors. If
vector is desired, use 1 row/col matrix.
}
\examples{
M <- diag(1:4)
N <- matrix(1:16, ncol = 4)

# Standard R syntax
M\%*\%N            

matrixProd(M, N) # using function

# Hadamard Product
M * N            
}
\seealso{
\code{\%*\%}
}
\author{
Stu Field
}
