% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixX.R
\name{matrixX}
\alias{matrixX}
\title{Matrix Multiplication}
\usage{
matrixX(A, B)
}
\arguments{
\item{A}{The pre-multiplied matrix.}

\item{B}{The post-multiplied matrix.}
}
\value{
A matrix (or vector) result of matrix multiplication.
}
\description{
Mimic of proper matrix multiplication without using the R syntax of %*%.
}
\details{
For matrix multiplication, the ncol(A) must equal nrow(B)! and you get a
nrow(A) x ncol(B) matrix as a result. Must be matrices not vectors. If
vector is desired, use 1 row/col matrix.
}
\note{
This is a mimic of the MatLab version of matrix multiplication, A * B,
where here is it is \code{matrixX(A, B)}.
}
\examples{

M <- diag(1:4)
N <- matrix(1:16, ncol = 4)
M\%*\%N            # R syntax
matrixX(M, N)    # using function
M * N            # Hadamard Product

}
\seealso{
\code{\%*\%}
}
\author{
Stu Field
}
