% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMetaData.R
\name{mergeMetaData}
\alias{mergeMetaData}
\title{Merge Meta Data}
\usage{
mergeMetaData(df, meta, ...)
}
\arguments{
\item{df}{The existing data frame to which the meta data is to be merged.}

\item{meta}{Can be either:
\enumerate{
\item A data frame of additional meta data, with at
least one common column name to match the samples OR
\item A file name (and path) pointing to where to get a
"*.csv" file containing meta data (this option uses a call
to \code{\link[utils]{read.csv}}).
}}

\item{...}{Additional arguments passed to \code{\link[dplyr]{left_join}}.}
}
\value{
A tibble object with extra columns corresponding to the meta data
added and indexed to the appropriate sample IDs.
}
\description{
Merge additional meta data to an existing data frame based on
unique row sample identifiers.
This is a wrapper around the existing R function
\code{\link[dplyr]{left_join}}.
}
\examples{

set.seed(101)
new <- data.frame(pid = sample(test_data$pid, 10), new.clin = rnorm(10))
mergeMetaData(test_data, new)            # use default 'by'
mergeMetaData(test_data, new, by ="pid") # same

}
\seealso{
\code{\link[dplyr]{left_join}}, \code{\link[utils]{read.csv}}
}
\author{
Stu Field
}
