% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton-raphson.R
\name{newton_raphson}
\alias{newton_raphson}
\title{Newton-Raphson Method}
\source{
\url{https://en.wikipedia.org/wiki/Newton\%27s_method}
}
\usage{
newton_raphson(init, tol = 1e-04)
}
\arguments{
\item{init}{The initial \code{x} value to begin Newton-Raphson steps.}

\item{tol}{The tolerance for the error in the root. Used to stop the
algorithm once the root estimate no longer changes by this much in \code{x}.}
}
\value{
The function generates a \pkg{ggplot2} of the algorithm, as well
as a list containing:
\item{x_traj}{The iterative "guesses" for the function root.}
\item{niter}{The number of iterations required to find the root.}
\item{root}{The estimate of the root (x-intercept).}
}
\description{
The Newton-Raphson Method is an iterative algorithm for calculating the root
(x-intercept) of an arbitrary function \code{f(x)}.
}
\details{
The idea is to start with an initial guess which is reasonably close to the
true root, then to approximate the function by its tangent line using
calculus, and finally to compute the x-intercept of this tangent line by
elementary algebra. This x-intercept will typically be a better
approximation to the original function's root than the first guess, and
the method can be iterated until some threshold tolerance is crossed.

Newton-Raphson Method for calculating next \code{x1} has the equation:

\deqn{
 x1 = x0 - f(x0) / f'(x0)
}
}
\examples{
newton_raphson(3)
newton_raphson(-3)
}
