% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormFun.R
\name{NormFun}
\alias{NormFun}
\title{Normal Distribution Histogram}
\usage{
NormFun(n = 1000, mu = 400, sd = 25, bks = 25, ...)
}
\arguments{
\item{n}{Numeric. The number of observations.}

\item{mu}{Numeric. A vector of means.}

\item{sd}{Numeric. A vector of standard deviations.}

\item{bks}{How to set the breaks for plotting the histogram.}

\item{...}{Additional arguments passed to \code{\link{hist}}.}
}
\value{
A histogram with the random observations produced in rnorm() with a
line produced with \code{\link[stats]{dnorm}}.
}
\description{
Describes a Gaussian (Normal) distribution given various
arguments and plots them in a histogram along with a line
composed by \code{\link[stats]{dnorm}}.
}
\details{
Neat function for making histogram with density line as well.
}
\note{
Used as an example to demo how to create functions in the R tutorial.
}
\examples{

NormFun()

}
\seealso{
\code{\link{hist}}, \code{\link{rnorm}}, \code{\link{dnorm}}, \code{\link{curve}}
}
\author{
Stu Field
}
