% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{objects}
\alias{objects}
\alias{TreeData}
\alias{popdata}
\alias{Ri_data}
\alias{MC_sims}
\alias{InfectionByCounty}
\alias{test_data}
\title{Package Objects}
\source{
TreeData: Stu Field

popdata: Stu Field

InfectionByCounty: Dylan George

test_data: Stu Field
}
\usage{
TreeData

popdata

Ri_data

MC_sims

InfectionByCounty

test_data
}
\description{
Below is a listing of all the objects contained in the
"stuRpkg" beginner package. There is a short description of
each object, and perhaps a description of its elements, how they were
derived, and their sources. Authorship of the object is attributed where
it is known.
}
\section{Functions}{
\itemize{
\item \code{TreeData}: Data set containing tree characteristic metric data,
that has been blinded
and anonymized from its original source. Used heavily in the \code{R}
tutorial and for general data frame usage (e.g. package testing).
The format is "tibble" data frame containing 20 cases and numerous random variables

\item \code{popdata}: A data set it the data produced by the
projection for the 6 class example used in the EEID R tutorial.
The format is a matrix of the 6-class population projection with classes
as columns and the map/step solutions as the rows

\item \code{Ri_data}: Coefficient Interclass Correlation Data Example
from Sokal & Rohlf (Biometry; 3rd ed.), pages 210-214.
The format is a data frame of data with treatments as
columns and rows as  cases.

\item \code{MC_sims}: A vector containing 1000 Monte-Carlo simulation
estimates of the integral estimate (area under curve) for a
given function. The simulation is time greedy so these
data are saved as an object.
The format is a numeric vector of length = 1000.

\item \code{InfectionByCounty}: Infection data describing the infections
count data of an infectious disease
by geographic position. Includes the following headings: \cr
\itemize{
\item CountyNo
\item CountyNo
\item CountyName
\item State
\item Fips
\item Site
\item Lat
\item Long
\item Population
\item Area
\item Density
\item Infected
}
A "tibble" data frame containing 3082 individual cases/records.

\item \code{test_data}: A quick sample tibble data frame for
running examples and checking data frame functionalities.
See \code{?tibble}.
\tabular{lr}{
Variable       \tab Definition \cr
pid            \tab seq(1041, 1080, 1) \cr
Pop            \tab rep(LETTERS[\verb{1:10]}) \cr
Sample         \tab sample(c("small", "medium", "large"), 20) \cr
TimePoint      \tab rep(c("baseline", "6 mths", "12 mths", "24 mths"), each = 10) \cr
a              \tab 1:length(test_data$TimePoint) \cr
b              \tab sample(\verb{1:10}, 40, replace = TRUE) \cr
ABCD.1234.56.8 \tab rnorm(40, 25, 3) \cr
XYZZ.6969.4.7  \tab rnorm(40, 25, 3) \cr
x              \tab test_data$a * runif (length(test_data$a), 0.25, 0.75) \cr
y              \tab sample(11:20, 40, replace = TRUE) \cr
z              \tab Mod(round(sample(rnorm(40)), 3)) \cr
Response       \tab sample(c("control", "disease"), 40, replace = TRUE)) \cr
}
}}

\examples{
head(TreeData)
summary(TreeData)
suppressWarnings(sapply(TreeData, mean))

head(popdata, 10)

Ri_data

head(MC_sims)
tail(MC_sims)
mean(MC_sims)
hist(MC_sims)

head(InfectionByCounty)

dim(test_data)
head(test_data)
}
\references{
\code{TreeData}: originally from Sala et al., but heavily modified to be
unrecognizable and used as the main example data set for the EEID R tutorial.

\code{popdata}: Stu Field (The EEID R Tutorial)

\code{Ri_data}: Sokal & Rohlf (Biometry; 3rd ed.), p. 210-214.

\code{MC_sims}: Bill Black IV

\code{InfectionByCounty}: originally from Dylan George and
used for an exercise data set in the EEID R tutorial (2010).
}
\keyword{Ri}
\keyword{Rtutorial}
