% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{optimizer}
\alias{optimizer}
\title{Optimization Function}
\usage{
optimizer(fn, I, max = TRUE, ...)
}
\arguments{
\item{fn}{Function to be evaluated. Must return a scalar value}

\item{I}{Numeric. Interval over which to find the max (or min)}

\item{max}{Logical. If FALSE, the Min is computed}

\item{...}{Arguments passed to the function}
}
\value{
A plot containing the functional curve, max/min, y-value at Max/Min
}
\description{
Given a function, determine the set of parameters that maximizes
(default) or minimizes it over a given interval via brute force.
}
\note{
This differs from the \code{\link{optim}} function but essentially
does the same. This is a brute force method.
}
\examples{
myfun <- function(x, a, b, t) {
   -t * (x-a)^2 + b
}

optimizer(fn = myfun, I = c(-10, 10), max = TRUE, a = 2, b = 12, t = 0.1)
optimizer(fn = myfun, I = c(-10, 10), max = FALSE, a = 2, b = 12, t = 0.1)
}
\seealso{
\code{\link{optimize}}, \code{\link{optim}}
}
\author{
Stu field
}
