% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2file.R
\name{path2file}
\alias{path2file}
\title{Determine Path to File}
\usage{
path2file(file, root.dir = Sys.getenv("HOME"), switchDir = FALSE)
}
\arguments{
\item{file}{Character string of the desired file, typically \code{"*.R"}
or \code{"*.csv"}}

\item{root.dir}{The path of the root directory. The higher
up the root to travel the longer the search will take
If you narrow down the path, more specific, it will take less
time. The default is a Mac/Unix formulation. For Windows/PC machines,
\code{root.dir} must be specified}

\item{switchDir}{Logical. Set the working directory to the path of the file in
\code{file}, if found. Uses \code{\link{setwd}}.}
}
\value{
The path of the file specified in \code{file}
}
\description{
Determine the path of a defined file via brute force search of root
directory and optionally change the R working directory to the directory
containing that file. Useful in setting the working directory prior to a
\code{\link{read.csv}} call.
}
\details{
If the file in \code{filename} is not found, function is
stopped and warning printed.
}
\examples{

\dontrun{
filename <- "TreeData.csv"
path2file(filename)
path2file(filename, switchDir = TRUE)
read.csv(filename)
}

}
\references{
Assistance from Steven Mosher via StackOverflow
}
\seealso{
\code{\link{list.files}}, \code{\link{setwd}}
}
\author{
Stu Field, Steven Mosher
}
