% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotErrorBars.R
\name{plotErrorBars}
\alias{plotErrorBars}
\title{Plot Error Bars}
\usage{
plotErrorBars(x, lo, up, plotfun = graphics::barplot, bar.col = 1,
  ylims = NULL, bar.lty = 1, bar.cex = 0.33, bar.lwd = 1,
  pad = 0.05, ...)
}
\arguments{
\item{x}{Summary data to plot into a barplot, typically a vector of the bar
heights or points.}

\item{lo}{Numeric. A vector of heights of the \emph{lower} error bars to be drawn.
Must be computed externally and passed into the function.}

\item{up}{Numeric. A vector of heights of the \emph{upper} error bars to be drawn.
Must be computed externally and passed into the function.}

\item{plotfun}{Currently one of: \code{barplot, points, or plot}. Note:
plot or points produce the same output.}

\item{bar.col}{Color of the error bars and lines.}

\item{ylims}{The y-axis limits to pass to \code{plotfun}.}

\item{bar.lty}{Line type for the error bars. Argument is passed eventually
to \code{\link[graphics]{lines}}, \code{options = 1:5}.}

\item{bar.cex}{Character expansion for the width of the error bars
(hat/base) in units of \emph{bar widths}.}

\item{bar.lwd}{Line width for the error bars (\code{lwd}) passed to
\code{\link[graphics]{lines}}.}

\item{pad}{Additional spacing, as a proportion of the y-value range,
to pad the y-axis limits.}

\item{...}{Additional arguments passed to either \code{\link[graphics]{barplot}}
, \code{\link[graphics]{plot}}, or \code{\link[graphics]{plot}}.}
}
\value{
A plot with error bars plotted on top defined by \code{up} and
\code{lo}.
}
\description{
Produce a barplot (or points) containing plot error bars.
}
\examples{
set.seed(101)
x <- rnorm(10)
plotErrorBars(x, lo = x - sd(x), up = x + sd(x))

# Barplot with Error Bars (SEMs & CIs)
set.seed(101)
f <- factor(rep(LETTERS[1:4], each = 20))
x <- runif(80)
data <- data.frame(x, f)
GroupMeans <- tapply(data$x, data$f, mean)
GroupSD <- tapply(data$x, data$f, sd)
sem <- GroupSD / sqrt(length(GroupMeans))
ci <- (GroupMeans + (sem * 1.96)) - (GroupMeans - (sem * 1.96))

par(mfrow=c(1, 2))
plotErrorBars(GroupMeans, lo = GroupMeans - sem, up = GroupMeans+sem)
plotErrorBars(GroupMeans, lo = GroupMeans - ci, up = GroupMeans+ci)

# Points with Error Bars
plotErrorBars(GroupMeans, lo = GroupMeans - sem, up = GroupMeans+sem, plotfun = plot,
              bar.col = 2, bar.lwd = 2, pch = 21, bg = 4, cex = 2)
}
\references{
Heavily modified from The R Book. 2007. Michael Crowley.
}
\seealso{
\code{\link[gplots]{plotCI}}, \code{\link[graphics]{barplot}}, \code{\link[graphics]{plot}}, \code{\link[graphics]{lines}}
}
\author{
Stu Field
}
