% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print2file.R
\name{print2file}
\alias{print2file}
\title{Print Object to File}
\usage{
print2file(..., file, width = 250)
}
\arguments{
\item{...}{Object(s) to be printed to file as standard output.}

\item{file}{File name for the output.}

\item{width}{Numeric. Controls the maximum number of columns on a line used
in printing vectors, matrices and arrays, and when filling by \code{\link[=cat]{cat()}}.
See also \code{getOption("width")}.}
}
\description{
A thin wrapper around the \code{\link[=sink]{sink()}} function to print
the corresponding object to a file connection.
}
\examples{
tab <- crossTab(mtcars, c("cyl", "carb"))
print2file(tab, file = "table_file.txt")
}
\seealso{
\code{\link[=sink]{sink()}}
}
\author{
Stu Field
}
