% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PunnettPack.R
\name{PunnettPack}
\alias{PunnettPack}
\title{Packing Punnett Map Solutions (FEScUE)}
\usage{
PunnettPack(x, nallele, nstages)
}
\arguments{
\item{x}{The object to be packed into an array. Typically a
vector of length = nallele^2 * nstages.}

\item{nallele}{Number of alleles}

\item{nstages}{Number of base stages.}
}
\value{
A 3-dim array with \code{dim = c(nallele, nallele, nstages)}.
}
\description{
Used to pack and unpack the "Punnett Square" formulation for
Tribolium model in FEScUE. Each slice of the 3D array is a
"punnett square" that is:
\cr
\deqn{ 
  \left(
  \begin{array}{cc}
    AA & Aa \\ 
    aA & aa 
  \end{array}
  \right)
}
}
\note{
Based on the Jesse Drendel formulation of the Tribolium model in
FEScUE for the Tribolium Project.
}
\examples{

PunnettPack(1:24, nallele=2, nstages=6)

}
\author{
Stu Field, Jesse Drendel
}
