% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeColumns.R
\name{removeColumns}
\alias{removeColumns}
\alias{rm.col}
\title{Remove Column Containing Given Entry(s)}
\usage{
removeColumns(x, index = 2, search)
}
\arguments{
\item{x}{A matrix or data frame}

\item{index}{Numeric (1,2). The margin to index (columns = 2; rows = 1)}

\item{search}{Numeric or character. The exact match to be searched.
If contained within any column or row, it will be removed}
}
\value{
Matrix or data frame with either row(s) or column(s) removed as
a result of matching \code{search}
}
\description{
Remove an entire column of a matrix of data frame whose
entries contain a given value. Created in SNP analysis data
and is similar to \code{\link{na.omit}} except the columns (default)
are removed is a match occurs in that margin.
}
\note{
Used in SNP analysis to remove nucleotide positions containing a given
character
}
\examples{

M <- matrix(1:25, ncol=5)
removeColumns(M, index=2, search=20)
M <- as.data.frame(M)
M[4, 4] <- "A"
M
removeColumns(M, search=c("A", 17, 23))

}
\seealso{
\code{\link{na.omit}}
}
\author{
Stu Field
}
