% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMat.R
\name{rMat}
\alias{rMat}
\title{Create Random Matrix}
\usage{
rMat(m, n, min, max, replace = FALSE, ...)
}
\arguments{
\item{m}{Numeric. Row dimension of the matrix to be produced.}

\item{n}{Numeric. Column dimension of the matrix to be produced.}

\item{min}{Numeric. Minimum random number in range.}

\item{max}{Numeric. Maximum random number in range.}

\item{replace}{Logical. Should the sampled numbers be repeated?}

\item{...}{Additional arguments passed to \code{\link{matrix}}.}
}
\value{
A matrix with m x n dimensions full of random numbers
on the interval \verb{[min, max]}.
}
\description{
Generate a matrix with random numbers in its entries.
}
\note{
Never know when you need a matrix of random numbers!
}
\examples{

rMat(5, 6, 35, 75)
rMat(5, 6, 35, 75, replace = TRUE)

}
\seealso{
\code{\link[base]{sample}}
}
\author{
Stu Field
}
