% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotateMatrix.R
\name{rotateMatrix}
\alias{rotateMatrix}
\title{Rotate a Matrix}
\usage{
rotateMatrix(x, direction = c("clock", "counter-clock"))
}
\arguments{
\item{x}{A matrix object.}

\item{direction}{Character. Which direction to rotate 90 degrees.
Partial matching allowed.}
}
\value{
A matrix rotated 90 degrees from the original matrix.
}
\description{
Function rotates a matrix 90 degrees. Direction can be either
clockwise or counter-clockwise.
}
\examples{
m <- matrix(1:9, ncol = 3)
m
rotateMatrix(m, "clock")
rotateMatrix(m, "cou")
}
\author{
Stu Field
}
