% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_replace.R
\name{search_replace}
\alias{search_replace}
\title{Search and Replace}
\usage{
search_replace(x, s, r)
}
\arguments{
\item{x}{The object to be searched, typically a matrix or data frame but
can be a vector of character or numeric class.}

\item{s}{The search index.}

\item{r}{The replace with. Must be same length as \code{s}.}
}
\value{
An object of the same dimensions and class as \code{x}, with the
\code{s} matches replaced with \code{r}.
}
\description{
A global search & replace of entries within a vector, matrix, or data frame
\strong{Note:} The lengths of \code{s} and \code{r} \emph{must} be identical.
}
\examples{
# matrix
Y <- matrix(1:25, ncol = 5)
Y
search_replace(Y, s = c(8, 20), r = c(99, 99))

# data.frame
X <- data.frame(x = c(1, 2, 3), y = c(2, 2, 4), z = c(1, 2, 4))
rownames(X) <- c("one", "two", "three")
X
search_replace(X, s = 1:4, r = c("A", "C", "G", "T"))

# numeric
search_replace(1:10, s = 4, r = 19)

# character
search_replace(head(LETTERS, 10), s = "G", r = "Z")
}
\author{
Stu Field
}
