% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchReplace.R
\name{searchReplace}
\alias{searchReplace}
\title{Search & Replace}
\usage{
searchReplace(x, s, r)
}
\arguments{
\item{x}{The object to be searched, typically a matrix or data frame but
can be a vector of character or numeric class.}

\item{s}{The search index.}

\item{r}{The replace with. Must be same length as \code{s}.}

\item{...}{Pass through to various S3 methods.}
}
\value{
An object of the same dimensions and class as \code{x}, with the
\code{s=} matches replaced with \code{r=}.
}
\description{
A global search & replace of entries within a vector, matrix, or data frame
}
\details{
\bold{Warning:} The lengths of \code{s} and \code{r} \emph{must} be identical.
}
\examples{

# matrix
Y <- matrix(1:25, ncol = 5)
Y
searchReplace(Y, s = c(8, 20), r = c(99, 99))

# data.frame
X <- data.frame(x = c(1, 2, 3), y = c(2, 2, 4), z = c(1, 2, 4))
rownames(X) <- c("one", "two", "three")
X
searchReplace(X, s = 1:4, r = c("A", "C", "G", "T"))

# numeric
searchReplace(1:10, s = 4, r = 19)

# character
searchReplace(head(LETTERS, 10), s = "G", r = "Z")

}
\seealso{
\code{\link{replace}}
}
\author{
Stu Field
}
