% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{stutils-deprecated}
\alias{stutils-deprecated}
\alias{collapse2df}
\alias{mapVector}
\alias{Mapply}
\alias{mergeMetaData}
\alias{optimizer}
\alias{getFileExt}
\title{Deprecated function(s) of the \pkg{stutils} package}
\arguments{
\item{...}{A simple pass-through to a replacement alternative if available.}
}
\description{
These functions are provided for compatibility with older versions of
\pkg{stuRpkg} package. They may eventually be completely
removed, so please re-code your scripts accordingly as soon as possible.
}
\details{
\tabular{rl}{
\code{collapse2df()}   \tab now use \code{do.call(rbind, ...)} \cr
\code{mapVector()}     \tab now use \code{dplyr::inner_join(x, hashmap)} (hashmap contains mappings) \cr
\code{Mapply()}        \tab now use \code{\link[purrr:map2]{purrr::pmap()}} \cr
\code{mergeMetaData()} \tab now use \code{\link[dplyr:mutate-joins]{dplyr::left_join()}} \cr
\code{getFileExt()}    \tab now use \code{\link[fs:path_file]{fs::path_ext()}} \cr
\code{optimizer()}     \tab no replacement \cr
}
}
\author{
Stu Field
}
