% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stutils-package.R
\docType{package}
\name{stutils-package}
\alias{stutils}
\alias{stutils-package}
\title{stutils: Stu's Utilities Package}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Small suite of useful functions and utilities
    used for data manipulation and coding in the R environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stufield/stutils}
}

}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com}

}
\keyword{internal}
