% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{seconds2time}
\alias{seconds2time}
\alias{time2seconds}
\title{Determine Time from Seconds (& vice versa)}
\usage{
seconds2time(x)

time2seconds(x)
}
\arguments{
\item{x}{Character or Numeric. Of the form \code{"hh:mm:ss.ss"}
with 2 decimal point precision on the seconds to convert to seconds.
If numeric, the seconds to convert to a character string of
the form \code{"hh:mm:ss.ss"}.}
}
\value{
Either number of seconds (\code{numeric} or the time format
(\code{string}) in "hh:mm:ss.ss".
}
\description{
\code{\link[=time2seconds]{time2seconds()}} and \code{\link[=seconds2time]{seconds2time()}} determine the
time (format="hh:mm:ss.ss") from the value in seconds or vice versa.
Two digit hour precision optional.
}
\examples{
seconds2time(159.72)

time2seconds("3:44:12.04")
time2seconds("15:44:12.04")

}
\seealso{
\code{\link[=strsplit]{strsplit()}}, \code{\link[=grep]{grep()}}
}
\author{
Stu Field
}
