% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TribMate.R
\name{TribMate}
\alias{TribMate}
\title{Tribolium Mating Subroutine}
\usage{
TribMate(Y, f)
}
\arguments{
\item{Y}{A matrix containing an intermediate population with the classes as
rows & genotypes as cols. Columns should be ordered AA, Aa, aa. Only the
final class mates.}

\item{f}{Integer representing the fecundity (i.e. egg production) of the
female mating partner. Number of offspring produced per reproductive adult.}
}
\value{
A matrix of identical dimensions to Y containing non-zero entries in
the first row only representing the newly produced offspring (eggs) in the
population. This matrix can then simply be added to the current, or
intermediate, solution to obtain the current solution.
}
\description{
A sub-routine used to determine the number of offspring produced by the
current population according to Hardy-Weinberg predictions. Used in
\code{Tribolium9()} where only the top class (adults) mates.
}
\references{
Tribolium Example from Caswell (2008); pg 71. Perturbation
analysis of nonlinear matrix populationn models. Demographic Research 18: 59-116
}
\author{
Stu Field, Jesse Drendel
}
