% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryNULL.R
\name{tryNULL}
\alias{tryNULL}
\title{Try Catch Returns NULL}
\usage{
tryNULL(expr, default = NULL, quiet = TRUE)
}
\arguments{
\item{expr}{An R expression to try.}

\item{default}{The value to return if an error is encountered.
By default, and according to the function name, \code{NULL}.
This option is included for the rare case when something other
than \code{NULL} is desired.}

\item{quiet}{Logical. Should the error message
(if encountered) be printed?}
}
\value{
Either the value of the expression in \code{expr} or,
if an error is encountered, the value in \code{default}.
}
\description{
A convenient wrapper to capture errors in an expression
returning \code{NULL} rather than error out.
This simplifies error catching for downstream functionality via
\code{\link[=is.null]{is.null()}}. See \code{\link[=try]{try()}} and \code{\link[=tryCatch]{tryCatch()}}.
}
\examples{
tryNULL(log("a"))
tryNULL(log("a"), quiet = FALSE)
tryNULL(log("a"), default = 0)
}
\references{
More or less "borrowed" from the \pkg{plyr} package.
}
\seealso{
\code{\link[=try]{try()}}, \code{\link[=tryCatch]{tryCatch()}}
}
\author{
Stu Field
}
