% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennWrapper.R
\name{venn_wrapper}
\alias{venn_wrapper}
\title{Venn Diagram Plotting Routine}
\usage{
venn_wrapper(
  x,
  main,
  filename = NULL,
  edge.col = "transparent",
  ...,
  colors = seq(length(x)),
  num.cex = 1,
  alpha = 0.25,
  label.col = "gray35",
  fontfamily = "sans",
  main.fontface = "bold",
  sub.fontface = "bold",
  main.col = 1,
  sub.col = "gray35",
  main.cex = 3,
  sub.cex = 1,
  margin = 0.05,
  cat.fontface = "bold",
  cat.cex = 2,
  cat.col = "black",
  cat.fontfamily = "sans",
  rotation.degree = 0
)
}
\arguments{
\item{x}{A named list of vectors containing strings to match intersections.}

\item{main}{Character giving the main title of the diagram}

\item{filename}{Filename for image output, or if NULL returns the grid object itself}

\item{edge.col}{describe}

\item{...}{Additional arguments passed to the underlying plotting routine
from \code{\link[VennDiagram:venn.diagram]{VennDiagram::venn.diagram()}}.}

\item{colors}{describe}

\item{num.cex}{describe}

\item{alpha}{describe}

\item{label.col}{describe}

\item{fontfamily}{describe}

\item{main.fontface}{Character giving the fontface (font style) of the main title}

\item{sub.fontface}{Character giving the fontface (font style) of the subtitle}

\item{main.col}{Character giving the colour of the main title}

\item{sub.col}{Character Colour of the subtitle}

\item{main.cex}{Number giving the cex (font size) of the main title}

\item{sub.cex}{Number giving the cex (font size) of the subtitle}

\item{margin}{describe}

\item{cat.fontface}{describe}

\item{cat.cex}{describe}

\item{cat.col}{describe}

\item{cat.fontfamily}{describe}

\item{rotation.degree}{Numeric (0-360). Rotation for the entire diagram.}
}
\value{
A \code{grob} class object which can be sent to a plotting device or
printed to file with the \verb{filename =} argument.
}
\description{
A wrapper function for plotting Venn diagrams comprising 2 -> 5
intersections.
}
\details{
\tabular{cccl}{
  \bold{Argument}       \tab \bold{Venn Sizes} \tab \bold{Class}    
  \tab \bold{Description}\cr
  \code{lwd}            \tab 1,2,3,4,5           \tab \var{numeric}   
  \tab Vector giving the width of each circle's circumference\cr
  \code{lty}            \tab 1,2,3,4,5           \tab \var{numeric}   
  \tab Vector giving the dash pattern of each circle's circumference\cr
  \code{col}            \tab 1,2,3,4,5           \tab \var{character} 
  \tab Vector giving the colour of each circle's circumference\cr
  \code{fill}           \tab 1,2,3,4,5           \tab \var{character} 
  \tab Vector giving the colour of each circle's area\cr
  \code{alpha}          \tab 1,2,3,4,5           \tab \var{numeric}   
  \tab Vector giving the alpha transparency of each circle's area\cr
  \code{label.col}      \tab 1,2,3,4,5           \tab \var{character} 
  \tab Vector giving the colour for each area label (length = 1/3/7/15 based on 
  set-number)\cr
  \code{cex}            \tab 1,2,3,4,5           \tab \var{numeric}   
  \tab Vector giving the size for each area label (length = 1/3/7/15 based on 
  set-number)\cr
  \code{fontface}       \tab 1,2,3,4,5           \tab \var{character} 
  \tab Vector giving the fontface for each area label (length = 1/3/7/15 based on 
  set-number)\cr
  \code{fontfamily}     \tab 1,2,3,4,5           \tab \var{character} 
  \tab Vector giving the fontfamily for each area label (length = 1/3/7/15 based on 
  set-number)\cr
  \code{cat.pos}        \tab 1,2,3,4,5           \tab \var{numeric}   
  \tab Vector giving the position (in degrees) of each category name along the circle, 
  with 0 at 12 o'clock\cr
  \code{cat.dist}       \tab 1,2,3,4,5           \tab \var{numeric}   
  \tab Vector giving the distance (in npc units) of each category name from the edge of 
  the circle (can be negative)\cr
  \code{cat.cex}        \tab 1,2,3,4,5           \tab \var{numeric}   
  \tab Vector giving the size for each category name\cr
  \code{cat.col}        \tab 1,2,3,4,5           \tab \var{character} 
  \tab Vector giving the colour for each category name\cr
  \code{cat.fontface}   \tab 1,2,3,4,5           \tab \var{character} 
  \tab Vector giving the fontface for each category name\cr
  \code{cat.fontfamily} \tab 1,2,3,4,5           \tab \var{character} 
  \tab Vector giving the fontfamily for each category name\cr
  \code{cat.just}       \tab 1,2,3,4,5           \tab \var{numeric}   
  \tab List (length = 1/2/3/4 based on set number) of Vectors of length 2 indicating 
  horizontal and vertical justification for each category name\cr
  \code{cat.default.pos}\tab 1,2,3               \tab \var{character} 
  \tab One of c('outer', 'text') to specify the default location of category names 
  (cat.pos and cat.dist are handled differently)\cr
  \code{cat.prompts}   \tab 2                   \tab \var{numeric}   
  \tab Boolean indicating whether to display help text on category name positioning 
  or not\cr
  \code{margin}         \tab 1,2,3,4,5           \tab \var{numeric}   
  \tab Number giving the amount of whitespace around the diagram in grid units\cr
  \code{rotation.degree}\tab 1,2,3,4,5           \tab \var{numeric}   
  \tab Number of degrees to rotate the entire diagram\cr
  \code{rotation.centre}\tab 1,2,3,4,5           \tab \var{numeric}   
  \tab Vector of length 2 indicating (x,y) of the rotation centre\cr
  \code{rotation}       \tab 3                   \tab \var{numeric}   
  \tab Number giving the clockwise rotation of a three-set Venn diagram (1, 2, or 3)\cr
  \code{reverse}        \tab 3                   \tab \var{logical}   
  \tab Reflect the three-set Venn diagram along its central vertical axis of symmetry.  
  Use in combination with \code{rotation} to generate all possible set orders\cr
  \code{euler.d}        \tab 2, 3                \tab \var{logical}   
  \tab Enable Euler diagrams for two-set and three-set Venn diagrams 
  (Venn Diagrams with moveable circles)\cr
  \code{scaled}         \tab 2, 3                \tab \var{logical}   
  \tab Enable scaling for two-set and certain three-set Euler diagrams. 
  (euler.d must be true to enable this)\cr
  \code{sep.dist}       \tab 2, 3                \tab \var{numeric}   
  \tab Controls the separation between distinct circles in certain two-set or three-set 
  Euler diagrams.\cr
  \code{offset}         \tab 2, 3                \tab \var{numeric}   
  \tab Number between 0 and 1 giving the amount to offset the smaller circle by in the 
  inclusion type of two-set Euler diagram and certain similar three-set Euler 
  diagrams.\cr
  \code{inverted}       \tab 2                   \tab \var{logical}   
  \tab Flip the two-set Venn diagram along its vertical axis (distinguished from 
  \code{reverse})\cr
  \code{ext.text}       \tab 2                   \tab \var{logical}   
  \tab Allow external text labels when areas are small\cr
  \code{ext.percent}    \tab 2                   \tab \var{numeric}   
  \tab A vector (length 3) indicating the proportion that a partial area has to be 
  smaller than to trigger external text placement.  The elements allow for individual 
  control of the areas in the order of the first area, second area and intersection 
  area.\cr
  \code{ext.pos}   \tab 2                   \tab \var{numeric}   
  \tab A vector (length 1 or 2) giving the positions (in degrees) of the external 
  area labels along the circles, with 0 (default) at 12 o'clock\cr
  \code{ext.line.lwd}   \tab 2                   \tab \var{numeric}   
  \tab Width of line connecting to \code{ext.text}\cr
  \code{ext.line.lty}   \tab 2                   \tab \var{numeric}   
  \tab The dash pattern of the lines connecting the external area labels to their 
  anchor points.\cr
  \code{ext.dist}       \tab 2                   \tab \var{numeric}   
  \tab Vector of length 1 or 2 indicating length of external line (use negative values 
  to shorten the line )\cr
  \code{ext.length}     \tab 2                   \tab \var{numeric}   
  \tab Vector of length 1 or 2 indicating the proportion of the external line that is 
  drawn from the anchor to the text\cr
  }
}
\note{
I stole this from the \pkg{VennDiagram} package.
}
\examples{
int_list <- lapply(1:3, function(...) sample(LETTERS[1:10], 6)) \%>\%
  purrr::set_names(c("Larry", "Curly", "Mo"))
venn_wrapper(int_list, num.cex = seq(0.5, 2, length = 7), cat.cex = c(1, 1.5, 2),
             main = "Title Here", sub = "Subtitle")
}
\references{
See the \pkg{VennDiagram} package.
}
\seealso{
\code{\link[VennDiagram:venn.diagram]{VennDiagram::venn.diagram()}}
}
\author{
Stu Field
}
