% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennWrapper.R
\name{vennWrapper}
\alias{vennWrapper}
\alias{venn_default}
\title{Venn Diagram Plotting Routine}
\usage{
vennWrapper(x, ..., edge.col = "transparent", colors = seq(length(x)),
  num.cex = 1, alpha = 0.25, label.col = "gray35",
  fontfamily = "sans", main.fontface = "bold", sub.fontface = "bold",
  main.col = 1, sub.col = "gray35", main.cex = 3, sub.cex = 1.5,
  margin = 0.01, cat.fontface = "bold", cat.cex = 2,
  cat.col = "black", cat.fontfamily = "sans", rotation.degree = 0,
  filename = NULL)
}
\arguments{
\item{x}{A named list of vectors containing strings to match intersections.}

\item{...}{Additional arguments passed to the internal
\code{venn_default}, which was stolen mostly from \code{\link{VennDiagram}}.}

\item{edge.col}{describe}

\item{colors}{describe}

\item{num.cex}{describe}

\item{alpha}{describe}

\item{label.col}{describe}

\item{fontfamily}{describe}

\item{main.fontface}{describe}

\item{sub.fontface}{describe}

\item{main.col}{describe}

\item{sub.col}{describe}

\item{main.cex}{describe}

\item{sub.cex}{describe}

\item{margin}{describe}

\item{cat.fontface}{describe}

\item{cat.cex}{describe}

\item{cat.col}{describe}

\item{cat.fontfamily}{describe}

\item{rotation.degree}{Numeric (0-360). Rotation for the entire diagram.}

\item{filename}{Character. Path and file name to which to print the figure.
Default of \code{NULL} or \code{NA} prints to the null device.}
}
\value{
A grob class object which can be plotting to a device of printed to
file if \code{filename} is not \code{NULL}.
}
\description{
A wrapper function for plotting Venn diagrams comprising 1 - 5
intersections.
}
\details{
More details \dots
}
\note{
I stole this ...
}
\examples{
int_list <- lapply(1:3, function(...) sample(LETTERS[1:10], 6)) \%>\%
  purrr::set_names(c("Larry", "Curly", "Mo"))
vennWrapper(int_list, num.cex = seq(0.5, 2, length = 7), cat.cex = c(1, 1.5, 2),
            main = "Title Here", sub = "Subtitle")
}
\references{
See the \code{\link{VennDiagram}} package.
}
\seealso{
\code{\link{VennDiagram}}, \code{\link{grid}}
}
\author{
Stu Field
}
\keyword{venn}
