% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichEntry.R
\name{whichEntry}
\alias{whichEntry}
\title{Which Entry(s)}
\usage{
whichEntry(x, match)
}
\arguments{
\item{x}{A \code{matrix} object to be searched.}

\item{match}{Index to be matched.}
}
\value{
A matrix indicating which entries agree with \code{match}. Rows are
the number of different match(es) and column 1 is the row of a positive
match and column 2 is the column of the positive match.
}
\description{
Determine which entry of a matrix matches a given value.
Similar functions exist for a vector, but when applied to a
matrix entries are converted to a vector of the matrix columns.
This is not ideal. This function provides \verb{row x column}
addressing of the entries which agree with the \verb{match =} argument.
}
\examples{
M <- matrix(1:25, ncol = 5)
M[5,1]  <- 17
M[4, 2] <- 17
whichEntry(M, 17)
}
\seealso{
\code{\link[=which]{which()}}
}
\author{
Stu Field
}
