% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeros.R
\name{zeros}
\alias{zeros}
\title{Diagonal Matrix of Zeros}
\usage{
zeros(x)
}
\arguments{
\item{x}{Numeric. The dimensions (m x n) for the square matrix to
be produced, or the length of the diagonal of the square matrix.}
}
\value{
A square matrix of zeros with \code{m x n} dimensions
where both equal \code{x}.
}
\description{
Produce a matrix of zeros of dimensions determined by the
diagonal argument. Mimic of the \code{MatLab} function of a similar name.
}
\note{
Mimic of MatLab, even though MatLab is bullshit.
}
\examples{
zeros(7)
zeros(15)
zeros(c(3, 5))
}
\seealso{
\code{\link{diag}}, \code{\link{diagR}}
}
\author{
Stu Field
}
