% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print2file.R
\name{print2file}
\alias{print2file}
\title{Print Object to File}
\usage{
print2file(..., file, width = 250)
}
\arguments{
\item{...}{Object(s) to be printed to file as standard output.}

\item{file}{File name for the output.}

\item{width}{Numeric. Controls the maximum number of columns on a line used
in printing vectors, matrices and arrays, and when filling by \code{cat}.
See also \code{getOption("width")}.}
}
\description{
A convenient wrapper around the \code{\link[base]{sink}} function to print
the resulting \code{\link[base]{print}} command to a corresponding file
connection. Useful for sharing data only results with collaborators or
medical group.
}
\examples{
tab <- crossTab(test_data, c("Pop", "Sample"))
print2file(tab, file = "table_file.txt")
}
\seealso{
\code{\link[base]{sink}}, \code{\link[base]{print}}
}
\author{
Stu Field
}
