% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuRpkg.R
\docType{package}
\name{stuRpkg-package}
\alias{stuRpkg}
\alias{stuRpkg-package}
\title{stuRpkg: The Stu R Package}
\description{
Small suite of useful functions, subroutines, and data objects for
  data set manipulation and analysis in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stufield/stuRpkg}
}

}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com}

}
\keyword{internal}
