% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryNULL.R
\name{tryNULL}
\alias{tryNULL}
\title{Try NULL Function}
\usage{
tryNULL(expr, default = NULL, quiet = TRUE)
}
\arguments{
\item{expr}{an R expression to try.}

\item{default}{The value to return if an error is encountered. By default,
and according to the function name, \code{NULL}. This option is included for
the rare case when something other than \code{NULL} is desired.}

\item{quiet}{Logical. Should the error message (if encountered) be printed
to the console?}
}
\value{
Either the value of the expression in \code{expr} or, if an error is
encountered, the value in \code{default}.
}
\description{
A convenient wrapper around either \code{\link{try}} or \code{\link{tryCatch}}.
However, rather than of returning the an error signal/warning, \code{NULL} is
returned if an error is encountered. This makes it easier to test the status
of the returned object later in the code with \code{\link{is.null}}.
}
\examples{
tryNULL(log("a"))
tryNULL(log("a"), quiet = FALSE)
tryNULL(log("a"), default = 0)
}
\references{
More or less stolen from the \code{plyr} package.
}
\seealso{
\code{\link{try}}, \code{\link{tryCatch}}
}
\author{
Stu Field
}
